refactor: Move section nav bar to its own file
